feat: enable two-way binding of [(dragulaModel)]
We stopped internally modifying arrays in
572c1baea13ae991a2c70fa91a4ae78b348c7e17.

This adds a (dragulaModelChange) event on the directive, which by
convention supports [(dragulaModel)]="model" two-way binding.

BREAKING CHANGE:

**`dragulaModel` no longer modifies arrays internally. You have to use
`[(dragulaModel)]="myModel"`** if you want to save the changes the library
makes. In most cases, this will be an easy change. However, if you were
relying on the arrays being modified directly, that will no longer
work.